requirements.txt
Django>=5.2.9
djangorestframework>=3.15
django-cors-headers>=4.0
openpyxl>=3.0
psycopg[binary]>=3.2